+++ brewlib/Time/FrameClock.cs
@@ -4,31 +4,31 @@ namespace BrewLib.Time;
 
 public interface FrameTimeSource : ReadOnlyTimeSource
 {
-    double Previous { get; }
-    double Elapsed { get; }
+    float Previous { get; }
+    float Elapsed { get; }
 
     event EventHandler Changed;
 }
 public class FrameClock : FrameTimeSource
 {
-    public double Current { get; private set; }
-    public double Previous { get; private set; }
+    public float Current { get; private set; }
+    public float Previous { get; private set; }
 
-    public double Elapsed => Current - Previous;
+    public float Elapsed => Current - Previous;
 
-    public double TimeFactor => 1;
+    public float TimeFactor => 1;
     public bool Playing => true;
 
     public event EventHandler Changed;
 
-    public void AdvanceFrame(double duration)
+    public void AdvanceFrame(float duration)
     {
         Previous = Current;
         Current += duration;
 
         if (Previous != Current) Changed?.Invoke(this, EventArgs.Empty);
     }
-    public void AdvanceFrameTo(double time)
+    public void AdvanceFrameTo(float time)
     {
         Previous = Current;
         Current = time;
